move modal into ui
