--- packages/lexical-playground/src/plugins/ExcalidrawPlugin/index.ts
@@ -1,55 +1,0 @@
-/**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- */
-import {useLexicalComposerContext} from '@lexical/react/LexicalComposerContext';
-import {$wrapNodeInElement} from '@lexical/utils';
-import {
-  $createParagraphNode,
-  $insertNodes,
-  $isRootOrShadowRoot,
-  COMMAND_PRIORITY_EDITOR,
-  createCommand,
-  LexicalCommand,
-} from 'lexical';
-import {useEffect} from 'react';
-
-import {
-  $createExcalidrawNode,
-  ExcalidrawNode,
-} from '../../nodes/ExcalidrawNode';
-
-export const INSERT_EXCALIDRAW_COMMAND: LexicalCommand<void> = createCommand(
-  'INSERT_EXCALIDRAW_COMMAND',
-);
-
-export default function ExcalidrawPlugin(): null {
-  const [editor] = useLexicalComposerContext();
-  useEffect(() => {
-    if (!editor.hasNodes([ExcalidrawNode])) {
-      throw new Error(
-        'ExcalidrawPlugin: ExcalidrawNode not registered on editor',
-      );
-    }
-
-    return editor.registerCommand(
-      INSERT_EXCALIDRAW_COMMAND,
-      () => {
-        const excalidrawNode = $createExcalidrawNode();
-
-        $insertNodes([excalidrawNode]);
-        if ($isRootOrShadowRoot(excalidrawNode.getParentOrThrow())) {
-          $wrapNodeInElement(excalidrawNode, $createParagraphNode).selectEnd();
-        }
-
-        return true;
-      },
-      COMMAND_PRIORITY_EDITOR,
-    );
-  }, [editor]);
-
-  return null;
-}
